Backend/requirements.txt
# Base
numpy

# OpenAI/Whisper
git+https://github.com/openai/whisper.git
ffmpeg-python
ffmpeg

# file
Pillow

# DRF
djangorestframework

django-cors-headers

torch